requirements.txt
fastapi
uvicorn
bcrypt
python-jose
python-multipart


